+++ eidos/eidos_rng.cpp
@@ -26,29 +26,29 @@
 
 gsl_rng *gEidos_rng = nullptr;
 int gEidos_random_bool_bit_counter = 0;
-unsigned long int gEidos_random_bool_bit_buffer = 0;
-int gEidos_rng_last_seed = 0;
+uint32_t gEidos_random_bool_bit_buffer = 0;
+unsigned long int gEidos_rng_last_seed = 0;				// unsigned long int is the type used by gsl_rng_set()
 
 
-int EidosGenerateSeedFromPIDAndTime(void)
+unsigned long int EidosGenerateSeedFromPIDAndTime(void)
 {
-	long pid = getpid();
+	unsigned long int pid = getpid();
 	time_t t;
 	
 	time(&t);
 	t += pid;
 	
-	return static_cast<int>(t);
+	return t;
 }
 
-void EidosInitializeRNGFromSeed(int p_seed)
+void EidosInitializeRNGFromSeed(unsigned long int p_seed)
 {
 	// Allocate the RNG if needed
 	if (!gEidos_rng)
 		gEidos_rng = gsl_rng_alloc(gsl_rng_taus2);
 	
 	// Then set the seed as requested
-	gsl_rng_set(gEidos_rng, static_cast<long>(p_seed));
+	gsl_rng_set(gEidos_rng, p_seed);
 	
 	// remember the seed as part of the RNG state
 	gEidos_rng_last_seed = p_seed;
